cloud_topics/batcher: move batcher into level_zero
