+++ src/v/cloud_topics/BUILD
@@ -2,8 +2,6 @@ load("//bazel:build.bzl", "redpanda_cc_library")
 
 package(default_visibility = [
     "//src/v/cloud_topics/batch_cache:__pkg__",
-    "//src/v/cloud_topics/batcher:__pkg__",
-    "//src/v/cloud_topics/batcher/tests:__pkg__",
     "//src/v/cloud_topics/core:__pkg__",
     "//src/v/cloud_topics/core/tests:__pkg__",
     "//src/v/cloud_topics/frontend:__pkg__",
@@ -156,9 +154,9 @@ redpanda_cc_library(
         "//src/v/cloud_topics:data_plane_api",
         "//src/v/cloud_topics:ephemeral_cluster_services",
         "//src/v/cloud_topics/batch_cache",
-        "//src/v/cloud_topics/batcher",
         "//src/v/cloud_topics/core:read_pipeline",
         "//src/v/cloud_topics/core:write_pipeline",
+        "//src/v/cloud_topics/level_zero/batcher",
         "//src/v/cloud_topics/level_zero/reconciler",
         "//src/v/cloud_topics/read_path:fetch_handler",
         "//src/v/cloud_topics/throttler",
