+++ src/v/cloud_topics/data_plane_impl.cc
@@ -13,10 +13,10 @@
 #include "base/outcome.h"
 #include "cloud_storage/cache_service.h"
 #include "cloud_topics/batch_cache/batch_cache.h"
-#include "cloud_topics/batcher/batcher.h"
 #include "cloud_topics/core/read_pipeline.h"
 #include "cloud_topics/core/write_pipeline.h"
 #include "cloud_topics/ephemeral_cluster_services.h"
+#include "cloud_topics/level_zero/batcher/batcher.h"
 #include "cloud_topics/level_zero/reconciler/reconciler.h"
 #include "cloud_topics/read_path/fetch_request_handler.h"
 #include "cloud_topics/throttler/throttler.h"
