+++ src/v/cloud_topics/level_zero/batcher/aggregator.cc
@@ -8,7 +8,7 @@
  * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_topics/batcher/aggregator.h"
+#include "cloud_topics/level_zero/batcher/aggregator.h"
 
 #include "cloud_topics/core/serializer.h"
 #include "cloud_topics/core/write_request.h"
