+++ src/v/cloud_topics/level_zero/batcher/aggregator.h
