+++ src/v/cloud_topics/level_zero/batcher/batcher.h
