+++ src/v/cloud_topics/level_zero/batcher/tests/BUILD
@@ -9,9 +9,9 @@ redpanda_cc_gtest(
     deps = [
         "//src/v/bytes",
         "//src/v/bytes:iostream",
-        "//src/v/cloud_topics/batcher:aggregator",
         "//src/v/cloud_topics/core:serializer",
         "//src/v/cloud_topics/core:write_request",
+        "//src/v/cloud_topics/level_zero/batcher:aggregator",
         "//src/v/container:chunked_circular_buffer",
         "//src/v/model",
         "//src/v/model/tests:random",
@@ -38,12 +38,12 @@ redpanda_cc_gtest(
         "//src/v/cloud_storage_clients",
         "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:object_utils",
-        "//src/v/cloud_topics/batcher",
-        "//src/v/cloud_topics/batcher:aggregator",
         "//src/v/cloud_topics/core:event_filter",
         "//src/v/cloud_topics/core:serializer",
         "//src/v/cloud_topics/core:write_pipeline",
         "//src/v/cloud_topics/core:write_request",
+        "//src/v/cloud_topics/level_zero/batcher",
+        "//src/v/cloud_topics/level_zero/batcher:aggregator",
         "//src/v/model",
         "//src/v/model/tests:random",
         "//src/v/random:generators",
