+++ src/v/cloud_topics/level_zero/batcher/tests/aggregator_test.cc
@@ -8,9 +8,9 @@
  * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_topics/batcher/aggregator.h"
 #include "cloud_topics/core/serializer.h"
 #include "cloud_topics/core/write_request.h"
+#include "cloud_topics/level_zero/batcher/aggregator.h"
 #include "model/namespace.h"
 #include "model/record.h"
 #include "model/tests/random_batch.h"
