+++ src/v/cloud_topics/level_zero/batcher/tests/remote_mock.h
